+++ spec/rspec/core/example_spec.rb
@@ -38,6 +38,12 @@ RSpec.describe RSpec::Core::Example, :parent_metadata => 'sample' do
       example = RSpec.describe.example
       expect(example.rerun_argument).to eq("#{RSpec::Core::Metadata.relative_path(__FILE__)}:#{__LINE__ - 1}")
     end
+
+    it "emits a deprecation warning when used" do
+      example = RSpec.describe.example
+      expect_deprecation_with_call_site(__FILE__, __LINE__ + 1, /rerun_argument/)
+      example.rerun_argument
+    end
   end
 
   describe "#update_inherited_metadata" do