+++ Changelog.md
@@ -13,6 +13,10 @@ Bug fixes:
 * Fix exception presenter when the root cause exception has nil backtrace.
   (Zinovyev Ivan, #2903)
 
+Deprecations:
+
+* Add RSpec 4 deprecation warnings. (Phil Pirozhkov, #2880)
+
 ### 3.10.1 / 2020-12-27
 [Full Changelog](http://github.com/rspec/rspec-core/compare/v3.10.0...v3.10.1)